+++ lib/presentation/pages/repo/components/readme_markdown.dart
@@ -15,7 +15,7 @@ import '../../../../domain/repositories/repo/entities/repo.dart';
 import '../../../../domain/repositories/repo/repo_repository.dart';
 import '../../../../utils/assets/assets.gen.dart';
 import '../../../../utils/logger.dart';
-import '../../../controllers/url_launcher.dart';
+import '../../../controllers/url_launcher_controller.dart';
 import '../../../widgets/async_value_handler.dart';
 
 /// ReadmeMarkdownのキャッシュマネージャープロバイダー
